+++ semaphores.h
@@ -2,6 +2,7 @@
 #define SEMAPHORE_H_
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #ifdef __unix__
 #include "mock/semaphores.h"
@@ -19,8 +20,8 @@ void os_semaphore_take(semaphore_t *sem);
 /** Takes a semaphore non-blocking */
 bool os_semaphore_try(semaphore_t *sem);
 
-/** Acquire the semaphore blocking with timeout [ms] */
-bool os_semaphore_try_timeout(semaphore_t *sem, float timeout);
+/** Acquire the semaphore blocking with timeout [us] */
+bool os_semaphore_try_timeout(semaphore_t *sem, uint32_t timeout);
 
 /** Releases (posts) a semaphore. */
 void os_semaphore_release(semaphore_t *sem);
